Redesigned the VmScaling features
- Changed the VerticalVmScaling, replacing the previous under and
  overload Predicates with two Function<Vm, Double> to define the lower
  and upper utilization thresholds. Each function receives a Vm and must
  return the threshold value it represents, as a percentage between 0 and 1 (where 1 is 100%).
  It was changed from Predicate to Function to enable the class to
  internally know which are the thresholds used.
  With the Predicate, the class could only know if the Vm
  was under or overloaded, but not the thresholds used, since
  a Predicate just returns a boolean value.
- A Predicate is used for HorizontalVmScaling objects yet,
  since they may perform multiple verifications of different
  resources, using different thresholds, to define
  if it's time to create new VMs to balance the load.
- The thresholds returned by the new threshold functions, defined for VerticalVmScaling
  interface, are also used to implement different ways of scaling a Vm resource vertically.
  The ResourceScalingType interface was introduced to allow:
  incremental up/down scaling when a Vm over/underload is detected; or
  instantaneous scaling the resource by automatically calculating its new
  size which will move the Vm from the under/overload state.
- Examples were updated to demonstrate how simple is to use these
  different types of resource scaling by just setting the
  resourceScalingType attribute of the VmScaling object.
  If no ResourceScalingType is set, the ResourceScalingGradual is defined
  as default.
- The VerticalVmScaling.scalingFactor is defined as a percentage
  for any kind of resource that can be scaled.
  Previously the scaling of CPU was defined as an absolute value,
  but a percentage value may be more effective, independent of the
  number of CPUs the VM has. For instance, adding 1 CPU to a VM
  that is overloaded and has 32 CPUs may not be enough.
  While adding 1% more CPUs is equal to 3.2 new CPUs.
  Since this value is rounded up, the actual new CPUs will be 4.

+++ cloudsim-plus-examples/src/main/java/org/cloudbus/cloudsim/examples/CloudSimExample0.java
@@ -134,7 +134,7 @@ public class CloudSimExample0 {
         final long   storage = 10000; // vm image size (MEGABYTE)
         final int    ram = 512; // vm memory (MEGABYTE)
         final long   bw = 1000; // vm bandwidth (Megabits/s)
-        final int    pesNumber = 2; // number of CPU cores
+        final long   pesNumber = 2; // number of CPU cores
 
         return new VmSimple(vmList.size(), mips, pesNumber)
                 .setBroker(broker)
