+++ cloudsim-plus-examples/src/main/java/org/cloudsimplus/examples/CustomUtilizationModelExample.java
@@ -185,7 +185,7 @@ public class CustomUtilizationModelExample {
         long length = 10000; //in Million Structions (MI)
         long fileSize = 300; //Size (in bytes) before execution
         long outputSize = 300; //Size (in bytes) after execution
-        int  numberOfCpuCores = vm.getNumberOfPes(); //cloudlet will use all the VM's CPU cores
+        long numberOfCpuCores = vm.getNumberOfPes(); //cloudlet will use all the VM's CPU cores
 
         //Defines how RAM and Bandwidth resources are used
         UtilizationModel utilizationFull = new UtilizationModelFull();
